backend/requirements.txt
fastapi
uvicorn
supabase
google-generativeai
langchain
langchain-google-genai
python-dotenv